Move BABA to dwpc.py. Update function names in BAAB function
